// src/models/CarCheck.ts
type CarCheck = {
	uuid: string;
	title: string;
	battery: string;
	light_headlamp: string;
	light_signal: string;
	light_tail: string;
	light_rear: string;
	oil: string;
	water: string;
	brake: string;
	air: string;
	gas: string;
	engine: string;
	tires: string;
	bearing_fr: string;
	bearing_fl: string;
	bearing_rr: string;
	bearing_rl: string;
	created_at: Date;
	updated_at: Date;
};

export default CarCheck;